Some more refactoring, starting  form controller
