+++ analysis_options.yaml
@@ -2,6 +2,8 @@ include: package:lint/analysis_options.yaml
 
 linter:
   rules:
+    avoid_function_literals_in_foreach_calls: false
+    avoid_positional_boolean_parameters: false
     sort_child_properties_last: false
     sort_pub_dependencies: false
     unsafe_html: false
