+++ lib/src/features/app/app.dart
@@ -1,0 +1,1 @@
+export 'view/app_view.dart';
