+++ lib/src/features/compose/compose.dart
@@ -1,0 +1,1 @@
+export 'view/compose_view.dart';
