+++ lib/src/features/compose/view/widgets/metadados_list.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import '../../misc/form_metadados.dart';
+import 'widgets.dart';
+
+class MetadadosList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<FormMetadados>(
+      builder: (context, formMetadados, _) {
+        return ListView.builder(
+          physics: const NeverScrollableScrollPhysics(),
+          shrinkWrap: true,
+          itemCount: formMetadados.metadados.length,
+          itemBuilder: (_, index) {
+            return MetadadosCard(metadado: formMetadados.metadados[index]);
+          },
+        );
+      },
+    );
+  }
+}
