--- lib/src/views/app/app.dart
@@ -1,1 +1,0 @@
-export 'app_view.dart';
