--- lib/src/views/browse/browse.dart
@@ -1,1 +1,0 @@
-export 'browse_view.dart';
