--- lib/src/views/browse/widgets/widgets.dart
@@ -1,1 +1,0 @@
-export 'classe_info.dart';
