--- lib/src/views/compose/compose.dart
@@ -1,1 +1,0 @@
-export 'compose_view.dart';
