--- lib/src/views/home/home.dart
@@ -1,3 +1,0 @@
-export 'home_controller.dart';
-export 'home_view.dart';
-export 'widgets/tree/treeview.dart' show TreeviewController;
