--- lib/src/views/views.dart
@@ -1,5 +1,0 @@
-export 'app/app.dart';
-export 'browse/browse.dart';
-export 'compose/compose.dart';
-export 'description/description.dart';
-export 'home/home.dart';
